data-fetching/requirements.txt
python-dotenv
psycopg2
black
isort
flake8
mypy
google-cloud-bigquery-storage
google-cloud-bigquery
db-dtypes
pandas
requests
types-requests
graphqlclient
pytest
freezegun